Move rpc response type to interface
