+++ crates/sovereign-sdk/rollup-interface/src/node/rpc/mod.rs
@@ -3,7 +3,7 @@
 
 use std::collections::BTreeMap;
 
-use alloy_primitives::{U32, U64};
+use alloy_primitives::{Bytes, B256, U32, U64};
 use block::L2BlockResponse;
 use risc0_zkp::core::digest::Digest;
 use serde::{Deserialize, Serialize};
@@ -227,6 +227,34 @@ pub struct LastVerifiedBatchProofResponse {
 
 /// The ZK proof generated by the [`ZkvmHost::run`] method to be served by rpc.
 pub type ProofRpcResponse = Vec<u8>;
+
+/// A single state diff entry. `value` is `None` when the key was deleted.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct StateDiffEntryResponse {
+    /// Raw storage key preimage
+    pub key: Bytes,
+    /// Codec-encoded value bytes, or `None` if the key was deleted
+    pub value: Option<Bytes>,
+}
+
+/// Response of `citrea_getStateDiffByBlockNumber`: the state diff produced by re-executing
+/// a single L2 block. Entries are deduplicated (last write wins) and sorted by key so that
+/// responses from different nodes are directly comparable.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct BlockStateDiffResponse {
+    /// The re-executed L2 block height
+    pub block_number: U64,
+    /// Hash of the re-executed block
+    pub block_hash: B256,
+    /// State root the block was executed against (root of block N-1)
+    pub pre_state_root: B256,
+    /// State root after re-execution; guaranteed to match the stored root of block N
+    pub post_state_root: B256,
+    /// The state diff produced by the block
+    pub state_diff: Vec<StateDiffEntryResponse>,
+}
 
 /// Workaround to serialize [u8; 32] with rpc_hex when the hash is optional
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
